refactor python code
refactor python code by ops
