--- deeplink_ext/easyllm_ops/flash_attention.py
@@ -1,19 +1,0 @@
-# Copyright (c) 2024, DeepLink.
-
-from deeplink_ext.internevo_ops.flash_attention import (
-    flash_attn_qkvpacked_func,
-    flash_attn_kvpacked_func,
-    flash_attn_func,
-    flash_attn_varlen_qkvpacked_func,
-    flash_attn_varlen_kvpacked_func,
-    flash_attn_varlen_func,
-)
-
-__all__ = [
-    "flash_attn_qkvpacked_func",
-    "flash_attn_kvpacked_func",
-    "flash_attn_func",
-    "flash_attn_varlen_qkvpacked_func",
-    "flash_attn_varlen_kvpacked_func",
-    "flash_attn_varlen_func",
-]
