--- deeplink_ext/easyllm_ops/flash_attention_fallback.py
@@ -1,20 +1,0 @@
-# Copyright (c) 2024, DeepLink.
-
-from deeplink_ext.internevo_ops.flash_attention_fallback import (
-    flash_attn_qkvpacked_func_torch,
-    flash_attn_kvpacked_func_torch,
-    flash_attn_func_torch,
-    flash_attn_varlen_qkvpacked_func_torch,
-    flash_attn_varlen_kvpacked_func_torch,
-    flash_attn_varlen_func_torch,
-)
-
-
-__all__ = [
-    "flash_attn_qkvpacked_func_torch",
-    "flash_attn_kvpacked_func_torch",
-    "flash_attn_func_torch",
-    "flash_attn_varlen_qkvpacked_func_torch",
-    "flash_attn_varlen_kvpacked_func_torch",
-    "flash_attn_varlen_func_torch",
-]
