--- deeplink_ext/internevo_ops/adamw.py
@@ -1,5 +1,0 @@
-# Copyright (c) 2024, DeepLink.
-
-from deeplink_ext.interntrain_ops.adamw import AdamW
-
-__all__ = ["AdamW"]
