--- deeplink_ext/internevo_ops/rms_norm.py
@@ -1,5 +1,0 @@
-# Copyright (c) 2024, DeepLink.
-
-from deeplink_ext.interntrain_ops import MixedFusedRMSNorm
-
-__all__ = ["MixedFusedRMSNorm"]
