--- deeplink_ext/internevo_ops/rms_norm_fallback.py
@@ -1,5 +1,0 @@
-# Copyright (c) 2024, DeepLink.
-
-from deeplink_ext.interntrain_ops.rms_norm_fallback import MixedRMSNormTorch
-
-__all__ = ["MixedRMSNormTorch"]
