+++ deeplink_ext/ops/rotary_embedding/interntrain_rotary_embedding.py
@@ -1,0 +1,11 @@
+# Copyright (c) 2024, DeepLink.
+
+from deeplink_ext.utils import PlatformType, deeplink_ext_get_platform_type
+
+platform_type = deeplink_ext_get_platform_type()
+if platform_type == PlatformType.TORCH_DIPU:
+    from ._rotary_embedding_dipu import ApplyRotaryEmb, ApplyRotaryEmbQKV_
+else:
+    raise ImportError
+
+__all__ = ["ApplyRotaryEmb", "ApplyRotaryEmbQKV_"]
